Move DetailInfo to admin's components
