+++ src/pages/Admin/components/DetailInfo/styles.module.scss
